Include additional determination of draft status
Due to inconsistencies in the data provided by the
leagues (esp. WHL and USHL) another determination
of players draft status was included.
Player birthdates and names are compared against a
list of players drafted in the previous three years.
This is a fuzzy comparison using the Levenshtein
string similarity criterion.
If a player can be found in this list he is
excluded from further processing.